pyproject.toml
[project]
name = "reviewer"
version = "0.1.0"
description = "AI-powered Python code reviewer"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.69.0",
    "streamlit>=1.55.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/reviewer"]

